# Contextful validity errors, null-byte specs fix