This pull request adds meaningful errors while decoding as well as proper null-byte encoding.

### Benchmarks

New is the pull request and old is `upstream/main`. Test performed on a laptop with an Intel Core i5-8250U.

```
name          old time/op    new time/op    delta
Encode-8         269ns ±14%     266ns ±17%     ~     (p=0.382 n=8+8)
EncodeTo-8      1.78µs ±13%    1.67µs ± 8%     ~     (p=0.130 n=8+8)
Decode-8        2.74µs ±12%    4.45µs ± 6%  +62.42%  (p=0.000 n=8+8) <-- more is slower
DecodeFrom-8    3.38µs ±16%    5.07µs ±10%  +50.13%  (p=0.000 n=8+8) <-- more is slower

name          old speed      new speed      delta
Encode-8      44.8MB/s ±15%  45.4MB/s ±15%     ~     (p=0.382 n=8+8)
EncodeTo-8     185MB/s ±14%   197MB/s ± 9%     ~     (p=0.130 n=8+8)
Decode-8       120MB/s ±11%    74MB/s ± 6%  -38.55%  (p=0.000 n=8+8) <-- less is slower
DecodeFrom-8  97.5MB/s ±14%  64.7MB/s ±10%  -33.71%  (p=0.000 n=8+8) <-- less is slower
```