@nihaals 2021-03-26
Based on GitHub Actions benchmarks, `BenchmarkDecodeFrom` is the main benchmark affected, the rest looks like random variance.

It would be nice to have some tests for invalid inputs, maybe even checking the position (although that might be overkill).