@diamondburned 2021-03-26
> maybe even checking the position

`type InvalidRuneError rune` can be feasibly changed to something else to add this, for example

```go
type InvalidRuneError struct {
    Rune     rune
    Position int
}
```