@nihaals 2021-03-26
> `type InvalidRuneError rune` can be feasibly changed to something else to add this, for example
> 
> ```go
> type InvalidRuneError struct {
>     Rune     rune
>     Position int
> }
> ```

I think this change would almost exclusively be used for tests, so I'm not too sure. I was originally thinking parsing the error string for the reported position and checking if it matches what is expected (hence me calling it overkill).