@diamondburned 2021-03-27
> add tests for invalid inputs

Do you have a set of reference invalid inputs?