@nihaals 2021-03-27
> > add tests for invalid inputs
> 
> Do you have a set of reference invalid inputs?

A few examples:
1. Decode: `💖✨✨✨,,,,👉👈test💖💖,👉👈❤️👉👈💖💖✨🥺👉👈💖💖✨🥺,👉👈` (non-bottom characters)
2. Decode: `💖✨✨✨,,,,👉test👈💖💖,👉👈❤️👉👈💖💖✨🥺👉👈💖💖✨🥺,👉👈` (non-bottom characters)
3. Decode: `💖✨✨✨,,,,👉💖👈💖💖,👉👈❤️👉👈💖💖✨🥺👉👈💖💖✨🥺,👉👈` (character in byte separator)
4. Decode: `💖✨✨✨,,,,👉test👈💖💖,👉👈❤️👉👈💖💖✨🥺👉👈💖💖✨🥺,👉👈` (character in byte separator)
5. Decode: `💖✨✨✨,,,,👉👈,💖💖👉👈❤️👉👈💖💖✨🥺👉👈💖💖✨🥺,👉👈` (characters not in order)
6. Decode: `💖✨✨✨,,,,👉👈💖💖,👉👈❤️💖💖✨🥺👉👈💖💖✨🥺,👉👈` (missing byte separator after null byte)